requirements.txt
pyjwt
cryptography
python-dotenv